requirements.txt
# Web Framework
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2

# Autenticacion y Seguridad
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9

# Indices Espaciales
Rtree==1.3.0

# Visualizacion
matplotlib==3.9.2

# Utilidades
python-dateutil==2.9.0
snowballstemmer==2.2.0
opencv-contrib-python>=4.9.0.80
scikit-learn>=1.3
librosa>=0.10
soundfile>=0.12
numpy>=1.24
scipy>=1.11
kagglehub>=0.2
pytest
kagglehub>=0.2